+++ Fagocity it/src/fagocity/model/Enemy.java
@@ -6,7 +6,7 @@ import fagocity.controller.EnemyController;
 
 public class Enemy extends Actor {
 		
-	public Enemy(double x, double y, double velX, double velY, double radius, Color color) {
+	public Enemy(int x, int y, double velX, double velY, int radius, Color color) {
 		super(x, y, velX, velY, radius, color);
 		actorController = new EnemyController(this);
 	}
